+++ src/utils/perf.js
